moved logging to skills
