+++ src/main/java/lvp/Main.java
@@ -5,13 +5,14 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.regex.Matcher;
+
+import lvp.skills.logging.LogLevel;
+import lvp.skills.logging.Logger;
+
 import java.net.URI;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
-
-import lvp.logging.LogLevel;
-import lvp.logging.Logger;
 
 public class Main {
     private record Config(Path path, String fileNamePattern, int port, LogLevel logLevel){}
