+++ src/main/java/lvp/skills/logging/LogDestination.java
@@ -1,4 +1,4 @@
-package lvp.logging;
+package lvp.skills.logging;
 
 public interface LogDestination {
     void log(String formattedMessage);
