+++ src/main/java/lvp/skills/logging/LogEntry.java
@@ -1,4 +1,4 @@
-package lvp.logging;
+package lvp.skills.logging;
 
 public record LogEntry(String time, LogLevel level, String message) {
 }
